support_bot/requirements.txt
# Бот поддержки @HelpBit_bot (можно использовать общий venv с VPN-ботом)
python-telegram-bot>=20.0
